Enhance Niagara sample with mesh shader support and buffer management
- Added a new shader file `NiagaraMeshlet.slang` for mesh shader functionality.
- Updated `FalcorSceneAdapter.cpp` to use 16-bit integers for triangle and vertex counts in meshlets, improving memory efficiency.
- Introduced `uploadSceneBuffers` method in `Niagara.cpp` to manage scene buffer uploads and ensure compatibility with Shader Model 6.5.
- Enhanced the user interface to display meshlet counts in the GUI, improving feedback during scene rendering.
- Updated CMake configuration to include the new shader file in the build process.

+++ Source/Samples/Niagara/FalcorSceneAdapter.cpp
@@ -90,17 +90,17 @@ static void appendMeshlet(NiagaraFormat::Geometry& result,
     NiagaraFormat::Meshlet m = {};
     m.dataOffset = uint32_t(dataOffset);
     m.baseVertex = baseVertex + minVertex;
-    m.triangleCount = (uint8_t)meshlet.triangle_count;
-    m.vertexCount = (uint8_t)meshlet.vertex_count;
+    m.triangleCount = (uint16_t)meshlet.triangle_count;
+    m.vertexCount = (uint16_t)meshlet.vertex_count;
     m.shortRefs = shortRefs ? 1u : 0u;
     m.center[0] = meshopt_quantizeHalf(bounds.center[0]);
     m.center[1] = meshopt_quantizeHalf(bounds.center[1]);
     m.center[2] = meshopt_quantizeHalf(bounds.center[2]);
     m.radius = meshopt_quantizeHalf(bounds.radius);
-    m.cone_axis[0] = bounds.cone_axis_s8[0];
-    m.cone_axis[1] = bounds.cone_axis_s8[1];
-    m.cone_axis[2] = bounds.cone_axis_s8[2];
-    m.cone_cutoff = bounds.cone_cutoff_s8;
+    m.cone_axis[0] = (int16_t)bounds.cone_axis_s8[0];
+    m.cone_axis[1] = (int16_t)bounds.cone_axis_s8[1];
+    m.cone_axis[2] = (int16_t)bounds.cone_axis_s8[2];
+    m.cone_cutoff = (int16_t)bounds.cone_cutoff_s8;
     result.meshlets.push_back(m);
 }
